+++ micro_manager/config.py
@@ -50,6 +50,7 @@ class Config:
         self._adaptivity = False
         self._adaptivity_type = ""
         self._data_for_adaptivity = dict()
+        self._adaptivity_mappings = []
         self._adaptivity_n = 1
         self._adaptivity_history_param = 0.5
         self._adaptivity_coarsening_constant = 0.5
@@ -327,6 +328,15 @@ class Config:
                 raise Exception("Adaptivity type can be either local or global.")
 
             self._logger.log_info_rank_zero("Adaptivity type: " + self._adaptivity_type)
+
+            try:
+                self._adaptivity_mappings = self._data["simulation_params"][
+                    "adaptivity_settings"
+                ]["mappings"]
+            except BaseException:
+                self._logger.log_info_rank_zero(
+                    "Adaptivity will not interpolate outputs, only use representatives."
+                )
 
             if self._data["simulation_params"]["adaptivity_settings"].get(
                 "lazy_initialization"
@@ -849,6 +859,17 @@ class Config:
             Either "local" or "global" depending on the type of adaptivity computation
         """
         return self._adaptivity_type
+
+    def get_adaptivity_mapping_configs(self):
+        """
+        Get the mapping configurations for the adaptivity interpolation scheme.
+
+        Returns
+        -------
+        adaptivity_mapping_configs : list
+            List of adaptivity mapping configurations.
+        """
+        return self._adaptivity_mappings
 
     def get_data_for_adaptivity(self):
         """